Try to make the navbar responsive
In this attempt, I added Javascript to make the navbar respond to the
click made on the .nav_toggle. But, I was not able to do that, mainly
because I haven't internalized the script, as I am yet to explore it.

However, the site is completely responsiveness, without any glitches,
except for the inactivity of the hamburger menu. Will try to fix it
in the next commit.